+++ apps/backend/src/routes/usersRoutes.ts
@@ -7,7 +7,7 @@ import { idParamSchema, searchQuerySchema } from "../schemas/common.schemas.js";
 import { validateRequest } from "../middleware/validate.js";
 import { NotFoundError } from "../errors/AppError.js";
 import { Response, NextFunction } from "express";
-import { deleteUser } from "../controllers/userController.js";
+import { deleteUser, getUsers, searchUsersFreeText } from "../controllers/userController.js";
 import updateAvatar from "../controllers/avatarController.js";
 import { avatarSchema } from "../schemas/avatar.schema.js";
 
@@ -15,69 +15,11 @@ const router = Router();
 
 // Endpoint for listing all users
 // Users with "user" role can see only other users, while "admin" can see both "dev" and "user"
-router.get(
-  "/",
-  authMiddleware,
-  async (req: AuthRequest, res: Response, next: NextFunction) => {
-    try {
-      if (req.user?.role === "user" || req.user?.role === "dev") {
-        const users = await UserModel.find({ role: "user" }).select(
-          "username createdAt"
-        );
-        res.json(users);
-      }
-      if (req.user?.role === "admin") {
-        const users = await UserModel.find({
-          role: { $in: ["dev", "user"] }
-        }).select("username role createdAt");
-        res.json(users);
-      }
-    } catch (err) {
-      next(err);
-    }
-  }
-);
+router.get("/", authMiddleware, getUsers);
 
-router.get(
-  "/search",
-  authMiddleware,
-  validateRequest({ query: searchQuerySchema }),
-  async (req: AuthRequest, res: Response, next: NextFunction) => {
-    try {
-      const { query } = req.validatedQuery as { query: string };
+router.get("/search", authMiddleware, validateRequest({ query: searchQuerySchema }), searchUsersFreeText);
 
-      const escapedQuery = query.replace(/[-\/\\^$*+?.()|[\]{}]/g, "\\$&");
-
-      if (req.user?.role === "user" || req.user?.role === "dev") {
-        const users = await UserModel.find({
-          role: "user",
-          $or: [
-            { username: { $regex: escapedQuery, $options: "i" } },
-            { email: { $regex: escapedQuery, $options: "i" } }
-          ]
-        }).select("username createdAt");
-        res.json(users);
-      }
-      if (req.user?.role === "admin") {
-        const users = await UserModel.find({
-          role: { $in: ["dev", "user"] },
-          $or: [
-            { username: { $regex: escapedQuery, $options: "i" } },
-            { email: { $regex: escapedQuery, $options: "i" } }
-          ]
-        }).select("username role createdAt");
-        res.json(users);
-      }
-    } catch (err) {
-      next(err);
-    }
-  }
-);
-
-router.get(
-  "/:id",
-  authMiddleware,
-  validateRequest({ params: idParamSchema }),
+router.get("/:id", authMiddleware, validateRequest({ params: idParamSchema }),
   async (req: AuthRequest, res: Response, next: NextFunction) => {
     try {
       const id = req.params.id as string;
